convert test to mocha
